refactor output contoller to use DateProvider type
