+++ apps/gp2-server/src/controllers/output.controller.ts
@@ -5,7 +5,7 @@ import {
   VALIDATION_ERROR_MESSAGE,
 } from '@asap-hub/model';
 import Boom from '@hapi/boom';
-import { OutputDataProvider } from '../data-providers/output.data-provider';
+import { OutputDataProvider } from '../data-providers/types';
 import { ExternalUserDataProvider } from '../data-providers/types/external-user.data-provider.type';
 
 export default class Outputs implements OutputController {
@@ -100,12 +100,9 @@ export default class Outputs implements OutputController {
       updatedBy: outputUpdateData.updatedBy,
     };
 
-    const outputId = await this.outputDataProvider.update(
-      id,
-      outputUpdateDataObject,
-    );
+    await this.outputDataProvider.update(id, outputUpdateDataObject);
 
-    return this.fetchById(outputId);
+    return this.fetchById(id);
   }
 
   private async validateOutput(
