+++ apps/gp2-server/src/data-providers/types/index.ts
@@ -2,6 +2,7 @@ export * from './asset.data-provider.type';
 export * from './contributing-cohort.data-provider.type';
 export * from './external-user.data-provider.type';
 export * from './news.data-provider.type';
+export * from './output.data-provider.type';
 export * from './page.data-provider.type';
 export * from './user.data-provider.type';
 export * from './working-group-network.data-provider.type';
