+++ apps/gp2-server/src/data-providers/types/output.data-provider.type.ts
@@ -1,0 +1,9 @@
+import { DataProvider, gp2 as gp2Model } from '@asap-hub/model';
+
+export type OutputDataProvider = DataProvider<
+  gp2Model.OutputDataObject,
+  gp2Model.FetchOutputOptions,
+  gp2Model.OutputCreateDataObject,
+  null,
+  gp2Model.OutputUpdateDataObject
+>;
