+++ apps/gp2-server/test/data-providers/output.data-provider.test.ts
@@ -80,7 +80,7 @@ describe('Outputs data provider', () => {
     });
 
     describe('Document Types', () => {
-      test('should throw when the document type is null', async () => {
+      test('should return when the document type is null', async () => {
         const squidexGraphqlResponse = getSquidexOutputGraphqlResponse();
         squidexGraphqlResponse.findOutputsContent!.flatData.documentType = null;
         squidexGraphqlClientMock.request.mockResolvedValueOnce(
@@ -174,14 +174,12 @@ describe('Outputs data provider', () => {
       },
     );
 
-    test('Should throw a Not Found error when the output is not found', async () => {
+    test('Should return null when the output is not found', async () => {
       squidexGraphqlClientMock.request.mockResolvedValueOnce({
         findOutputsContent: null,
       });
 
-      await expect(outputDataProvider.fetchById(outputId)).rejects.toThrow(
-        'Not Found',
-      );
+      expect(await outputDataProvider.fetchById('not-found')).toBeNull();
     });
 
     test('Should throw an error with a specific error message when the graphql client throws one', async () => {
@@ -723,7 +721,7 @@ describe('Outputs data provider', () => {
     describe('Update', () => {
       const outputId = 'updated-output-id';
 
-      test('Should update the existing output and return its ID', async () => {
+      test('Should update the existing output', async () => {
         const outputUpdateData = getOutputUpdateDataObject();
 
         const restOutputUpdateData = getRestOutputUpdateData();
@@ -734,11 +732,8 @@ describe('Outputs data provider', () => {
           })
           .reply(201, { id: outputId });
 
-        const result = await outputDataProvider.update(
-          outputId,
-          outputUpdateData,
-        );
-        expect(result).toEqual(outputId);
+        await outputDataProvider.update(outputId, outputUpdateData);
+        expect(nock.isDone()).toBe(true);
       });
 
       test('Should throw when fails to update the output - 400', async () => {
